// types/operator-type.ts
export const arithmeticHeaders: string[] = ["Operator", "Operation", "Example", "Description"];
export interface arithmeticRows {
  operator: string;
  opertion: string;
  example: string;
  description: string;
}

export const assigmentHeaders: string[] = ["Operator", "Operation", "Example"];
export interface assigmentRows {
  operator: string;
  opertion: string;
  example: string;
}

export const comparationHeaders: string[] = ["Operator", "Meaning", "Example"];
export interface comparationRows {
  operator: string;
  meaning: string;
  example: string;
}

export const logicalHeaders: string[] = ["Operator", "Example", "Meaning"];
export interface logicalRows {
  operator: string;
  example: string;
  meaning: string;
}

export const bitwiseHeaders: string[] = ["Operator", "Meaning", "Example", "Description"];
export interface bitwiseRows {
  operator: string;
  meaning: string;
  example: string;
  description: string;
}

export const specialHeader: string[] = ["Operator", "Meaning", "Example"];
export interface specialRows {
  operator: string;
  meaning: string;
  example: string;
}

export const memberShipHeader: string[] = ["Operator", "Meaning", "Example"];
export interface memberShipRows {
  operator: string;
  meaning: string;
  example: string;
}